requirements.txt
# ===========================================
# Dingding Machine v2.0 依赖清单
# ===========================================

# 钉钉集成
dingtalk-stream==0.24.3
alibabacloud_dingtalk==2.2.39
alibabacloud_tea_openapi==0.4.3
alibabacloud_tea_util==0.3.14

# 任务调度
apscheduler==3.11.2

# 数据库驱动（多数据源支持）
cx_Oracle==8.3.0          # Oracle（旧版，保留兼容）
oracledb==2.0.0           # Oracle（新版推荐）
asyncpg==0.29.0           # PostgreSQL（异步）
mysql-connector-python==8.3.0  # Doris（MySQL协议）

# 数据处理
pandas==1.5.3
numpy==1.26.4
openpyxl==3.1.5
xlrd==2.0.2
xlutils==2.0.0
xlwt==1.3.0

# 配置和验证
PyYAML==6.0
pydantic==2.8.0
pydantic-settings==2.3.0

# 网络请求
aiohttp==3.13.2
Requests==2.33.1
httpx==0.27.0
websockets==15.0.1

# 安全
sqlparse==0.5.0

# 工具
chinese_calendar==1.11.0
setuptools==78.1.1

# ===========================================
# Agent 相关依赖（第二阶段需要）
# ===========================================
# langchain==0.3.0
# langgraph==0.2.0
# langchain-openai==0.2.0
# langchain-community==0.3.0

# ===========================================
# 可选依赖
# ===========================================
# redis==5.0.0            # 缓存（可选）
# structlog==24.4.0       # 结构化日志（可选）
# langsmith==0.1.0        # 可观测性（可选）
